Restructured code
Restructure code for better usability and understanding. No syntax changes for the user.
